consumer/requirements.txt
SQLAlchemy==1.4.51
psycopg2-binary==2.9.9
faust-streaming==0.10.22
